added ffence and ground

+++ Joint Project 1/Player.h
@@ -5,8 +5,8 @@ class Player
 {
 	// private data members
 	int speed = 20;
-	const float SCREEN_WIDTH = 1500;   // the size of the screen in pixels used in the game
-	const float SCREEN_HEIGHT = 800;
+	const int SCREEN_WIDTH = 1500;   // the size of the screen in pixels used in the game
+	const int SCREEN_HEIGHT = 800;
 	const int SPRITE_HEIGHT = 64;
 	const int SPRITE_WIDTH = 48;
 public:
